req service

+++ leave-back/internal/request-service/dto/dto.go
@@ -10,8 +10,8 @@ type CreateRequest struct {
 }
 
 type CreateLeaveType struct {
-	Name  string `json:"name" binding:"required"`
-	Quota int    `json:"quota" binding:"required"`
+	Name  string  `json:"name" binding:"required"`
+	Quota float64 `json:"quota" binding:"required"`
 }
 
 type CreateHolidays struct {
