+++ leave-back/shared/model/leave_types.go
@@ -1,9 +1,9 @@
 package model
 
 type LeaveTypes struct {
-	ID    int    `gorm:"column:id" json:"id"`
-	Name  string `gorm:"column:name" json:"name"`
-	Quota int    `gorm:"column:quota" json:"quota"`
+	ID    int     `gorm:"column:id" json:"id"`
+	Name  string  `gorm:"column:name" json:"name"`
+	Quota float64 `gorm:"column:quota" json:"quota"`
 }
 
 func (LeaveTypes) TableName() string {
